backend/requirements.txt
fastapi
uvicorn
python-multipart
pymongo
pydantic
requests
numpy
pinecone
pypdf
python-dotenv